feat: extract shared ActivityFeedItem component, sync activity rendering across UserProfile, Dashboard, and Activity
